// components/ui/types.ts
export interface Option {
	label?: string;
	description?: string;
	icon?: ReactNode;
	disabled?: boolean;
}